[ci] add minimal tests for docker images
closes #83

This commit add a minimal test suite that validates that the images,
that we build and publish, include a minimally working Theia application.

Until now we relied on the success of an image's build during CI, to
publish it on dockerhub. That's not good, because Theia applications
can sometimes build but not behave correctly or even start.

The UI test suite and related configurations were borrowed from the
main theia repo and slightly adapted. ATM we use only the most basic
test suite, "main-page", that should be applicable to all Theia
applications we run CI on.

This will be enhanced in the future to to add image-specific test
files. For example, see wdio.theia-docker.conf.js, that adds an
extra test file to be run, for the theia-docker image. ATM this second
suite is bogus and just for demonstration purposes.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>