Update travis config for repo, remove "sudo"
It seems that using "sudo" is no longer required for our config, since
we can use the default docker-ce.

Also it seems that when we use "travis_wait" with a docker build, the
build's output is not available in real-time, on Travis, but rather
the bulk of the logs are only added after a given build is done, and
then may be incomplete. A work-around was added to address that.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>