Revert "[temporary] Disable publish to dockerhub"
This reverts commit 31872398ae52dac2c08318f6d30b157baad7dcf9.

Now that all images we publish have been updated to use two stages,
we are able to re-enable publishing to dockerhub.

The reason for the update is that we realized that build arguments
are saved in the image's history, even if not present in the final
image as such. This means the GH token, used during the build,
was present for anyone to see, in the history of images we were
publishing. The old tokens we use have been revoked, and from now-on
the Theia app is built in a first stage that's discarded from the
final, published image.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>